feat: refactoring components and folders, routes/controllers in backend
