+++ src/Header.tsx
@@ -1,0 +1,7 @@
+const Header = () => (
+  <h1 className="text-3xl font-bold text-transparent bg-clip-text bg-gradient-to-r from-blue-400 to-purple-500">
+    CARDANO TRANSACTION
+  </h1>
+);
+
+export default Header;
